Eight cleanups (R2.1, R2.3, R2.5-R2.7, R3.1, R3.3-R3.5, R1.5, R1.3), node
count unchanged at 56158265. Two independent A/B timing runs measured
-0.15 % and -0.19 %, so the set is not slower.

+++ search/moveprovider.h
@@ -386,11 +386,16 @@ namespace QaplaSearch {
 		int16_t selectProposedMove(Move move) {
 			int16_t foundMoveNo = -1;
 			if (!move.isEmpty()) {
-				for (uint8_t moveNo = 0; moveNo < moveList.getTotalMoveAmount(); moveNo++) {
+				// uint32_t, not uint8_t: the generator produces up to MAX_MOVE_AMOUNT moves and
+				// an index that wraps at 255 never reaches the end of the list.
+				// No break on a hit: every move is generated once, so the first hit is the only
+				// one and stopping there is free work saved - but it measured 0.5 % slower in
+				// five interleaved pairs, consistently. The loop without an early exit is the
+				// one the compiler can lay out well, so the scan runs to the end on purpose.
+				for (uint32_t moveNo = 0; moveNo < moveList.getTotalMoveAmount(); moveNo++) {
 					if (moveList[moveNo] == move) {
-						foundMoveNo = moveNo;
+						foundMoveNo = int16_t(moveNo);
 					}
-
 				}
 			}
 			return foundMoveNo;
